requirements.txt
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.7.1
appnope==0.1.3
apprise==1.4.5
APScheduler==3.10.4
asgi-lifespan==2.1.0
asttokens==2.2.1
asyncpg==0.28.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
boto3==1.28.39
botocore==1.31.39
build==0.10.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.4
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
coolname==2.2.0
crashtest==0.4.1
croniter==1.4.1
cryptography==41.0.2
cycler==0.11.0
Cython==3.0.2
databricks-cli==0.17.7
dateparser==1.1.8
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
docker==6.1.3
dulwich==0.21.5
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
filelock==3.12.3
Flask==2.3.3
fonttools==4.42.1
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.33
google-auth==2.22.0
greenlet==2.0.2
griffe==0.31.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keyring==24.2.0
kiwisolver==1.4.5
kubernetes==26.1.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==2.6.0
more-itertools==10.1.0
msgpack==1.0.5
nest-asyncio==1.5.6
numpy==1.25.1
oauthlib==3.2.2
orjson==3.9.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
prefect==2.10.21
prefect-sqlalchemy==0.2.4
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
pyperclip==1.8.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python-telegram-bot==20.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.1.0
querystring-parser==1.2.4
rapidfuzz==2.15.1
readchar==4.0.5
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
scikit-learn==1.3.0
scipy==1.11.1
shellingham==1.5.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
trove-classifiers==2023.8.7
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.24.4
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.7
xattr==0.10.1
zipp==3.16.2
